requirements.txt
streamlit>=1.30.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
Pillow>=10.0.0
pandas>=2.0.0
plotly>=5.18.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface-hub>=0.25.0
tavily-python>=0.3.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.40.0
scipy>=1.10.0
tiktoken>=0.5.0
httpx>=0.25.0
